requirements.txt
openai
python-dotenv
langchain
transformers
requests
streamlit
